Add debug extension for site errors
This is intended to be an easy way to expose code errors that the
developer should be made aware of.  Added an extension to ‘debug’
(called ‘site’) that adds the debug message to the namespace
‘autohost:site’, regardless of the namespace of the debug object.  This
allows simplified debugging - simple set DEBUG=autohost:site in a
project’s gulpfile, and errors detected with site code will be output
in console.